+++ Runtime/RoadNetwork/Util/Vector2d.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 79ff0cb4b36570d4e984a1095a3c046a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
